Correct the tests for Pandas 0.23
