+++ python/pyarrow/tests/test_pandas.py
@@ -4082,6 +4082,18 @@ def test_array_to_pandas():
         # tm.assert_series_equal(result, expected)
 
 
+def test_roundtrip_empty_table_with_intervalrange_index():
+    if Version(pd.__version__) < Version("1.0.0"):
+        pytest.skip("ExtensionDtype to_pandas method missing")
+
+    df = pd.DataFrame(index=pd.interval_range(start=0, end=3))
+    table = pa.table(df)
+    table.to_pandas().index == pd.Index([{'left': 0, 'right': 1},
+                                         {'left': 1, 'right': 2},
+                                         {'left': 2, 'right': 3}],
+                                        dtype='object')
+
+
 # ----------------------------------------------------------------------
 # Legacy metadata compatibility tests
